moves test files into their own folder
