+++ src/tests/graphql-client.test.ts
@@ -1,5 +1,4 @@
-import { graphqlRequest } from './graphql-client';
-
+import { graphqlRequest } from '../lib/graphql-client';
 describe('graphqlRequest error handling', () => {
   it('throws error for invalid request object', async () => {
     await expect(graphqlRequest(null as any)).rejects.toThrow();
